Move PageWrapper and Logo to Shared folder
