+++ identity/webapp/src/frontend/Registration/AccountCreated.tsx
@@ -1,7 +1,7 @@
 import Icon from '@weco/common/views/components/Icon/Icon';
 import SpacingComponent from '@weco/common/views/components/SpacingComponent/SpacingComponent';
 import React from 'react';
-import { PageWrapper } from './PageWrapper';
+import { PageWrapper } from '../Shared/PageWrapper';
 import { Container, SuccessMessage, Wrapper } from './Registration.style';
 
 export const AccountCreated: React.FC = () => {
