+++ identity/webapp/src/frontend/Shared/OldPageWrapper.tsx
@@ -1,0 +1,8 @@
+import styled from 'styled-components';
+
+export const OldPageWrapper = styled.div`
+  width: 70%;
+  margin: auto;
+  background-color: #f0ede3;
+  padding: 42px;
+`;
